+++ front/scripts/main/components/LoginIconComponent.ts
@@ -8,10 +8,34 @@ App.LoginIconComponent = Em.Component.extend({
 	classNames: ['external', 'login'],
 
 	click: function (): void {
+		var label: string,
+			href: string;
+
 		if (Mercury.wiki.enableNewAuth) {
-			window.location.href = '/join?redirect=' + encodeURIComponent(window.location.href);
+			href = '/join?redirect=' +
+			encodeURIComponent(window.location.href) +
+			this.getUselangParam();
+			label = 'join-link';
 		} else {
-			window.location.href = '/Special:UserLogin';
+			label = 'legacy-login-link';
+			href = '/Special:UserLogin';
 		}
+
+		M.track({
+			trackingMethod: 'ga',
+			action: M.trackActions.click,
+			category: 'user-login-mobile',
+			label: label
+		});
+
+		window.location.href = href;
+	},
+
+	getUselangParam: function (): string {
+		var lang: string = Mercury.wiki.language.content;
+		if (!lang || lang === 'en') {
+			return ''
+		}
+		return '&uselang=' + encodeURIComponent(lang);
 	}
 });