+++ front/scripts/main/models/EditModel.ts
@@ -1,5 +1,6 @@
 /// <reference path="../app.ts" />
 /// <reference path="../../baseline/mercury" />
+/// <reference path="../mixins/EditMixin.ts" />
 'use strict';
 
 App.EditModel = Em.Object.extend({
@@ -13,44 +14,13 @@ App.EditModel = Em.Object.extend({
 	})
 });
 
-App.EditModel.reopenClass({
-
-	getEditToken: function(title: string): Em.RSVP.Promise {
-		return new Em.RSVP.Promise((resolve: Function, reject: Function): void => {
-			Em.$.ajax({
-				url: '/api.php',
-				data: {
-					action: 'query',
-					prop: 'info',
-					titles: title,
-					intoken: 'edit',
-					format: 'json'
-				},
-				dataType: 'json',
-				success: (resp: any): void => {
-					var edittoken: string,
-						pages: any = Em.get(resp, 'query.pages');
-					if (pages) {
-						// FIXME: MediaWiki API, seriously?
-						edittoken = pages[Object.keys(pages)[0]].edittoken;
-						resolve(edittoken);
-					} else {
-						reject();
-					}
-				},
-				error: (err): void => {
-					reject(err);
-				}
-			});
-		});
-	},
-
+App.EditModel.reopenClass(App.EditMixin, {
 	publish: function(model: any): Em.RSVP.Promise {
 		return new Em.RSVP.Promise((resolve: Function, reject: Function): void => {
 			this.getEditToken(model.title)
 				.then((token: any): void => {
 					Em.$.ajax({
-						url: '/api.php',
+						url: M.buildUrl({path: '/api.php'}),
 						data: {
 							action: 'edit',
 							title: model.title,
@@ -82,7 +52,7 @@ App.EditModel.reopenClass({
 
 	load: function(title: string, sectionIndex: number): Em.RSVP.Promise {
 		return new Em.RSVP.Promise((resolve: Function, reject: Function): void => {
-			Em.$.ajax('/api.php', {
+			Em.$.ajax(M.buildUrl({path: '/api.php'}), {
 				dataType: 'json',
 				cache: false,
 				data: {