+++ front/scripts/main/router.ts
@@ -13,14 +13,26 @@ App.Router.map(function () {
 
 	this.route('mainPage', {
 		path: '/'
+	}, function() {
+		this.route('section', {
+			path: '/main/section/:sectionName'
+		});
+
+		this.route('category', {
+			path: '/main/category/:categoryName'
+		});
 	});
 
 	this.route('article', {
-		path: articlePath + '*title'
+		path: articlePath + ':title'
 	});
 
 	this.route('edit', { // Symbolic link to EditController
 		path: articlePath + 'edit/:title/:sectionIndex'
+	});
+
+	this.route('addPhoto', { // Symbolic link to AddPhotoController
+		path: articlePath + 'addPhoto/:title'
 	});
 
 	// We don't want to duplicate the previous route