+++ server/facets/api/mainPageCategory.ts
@@ -7,18 +7,6 @@ import Utils = require('../../lib/Utils');
 import localSettings = require('../../../config/localSettings');
 import wrapResult = require('./presenters/wrapResult');
 
-interface CuratedContentSectionMW {
-	items: {
-		title: string;
-		label: string;
-		image_id: string;
-		article_id: string;
-		type: string;
-		image_url: string;
-		article_local_url: string
-	};
-}
-
 var cachingTimes = {
 	enabled: false,
 	cachingPolicy: Caching.Policy.Private,
@@ -29,20 +17,25 @@ var cachingTimes = {
 export function get (request: Hapi.Request, reply: any): void {
 	var params = {
 			wikiDomain: Utils.getCachedWikiDomainName(localSettings, request.headers.host),
-			sectionName: request.params.sectionName || null
+			categoryName: decodeURIComponent(request.params.categoryName),
+			thumbSize: request.params.thumbSize || {
+				width: 300,
+				height: 300
+			},
+			offset: request.query.offset || ''
 		};
 
-	if (params.sectionName === null) {
-		// TODO: ad hoc error handling, use Boom everywhere?
-		reply(Boom.badRequest('Section not provided'));
+	if (params.categoryName === null) {
+		reply(Boom.badRequest('Category not provided'));
 	} else {
-		new MW.ArticleRequest(params).curatedContentSection(params.sectionName)
-		.then((response: any): void => {
-			reply(response);
-			Caching.setResponseCaching(response, cachingTimes);
-		}, (error: any): void => {
-			var wrappedResult = wrapResult(error, {});
-			reply(wrappedResult).code(wrappedResult.status);
-		});
+		new MW.ArticleRequest(params)
+			.category(params.categoryName, params.thumbSize, params.offset)
+			.then((response: any): void => {
+				reply(response);
+				Caching.setResponseCaching(response, cachingTimes);
+			}, (error: any): void => {
+				var preparedResult: any = wrapResult(error, {});
+				reply(preparedResult).code(preparedResult.status);
+			});
 	}
 }