+++ CHANGELOG.md
@@ -1,3 +1,106 @@
+## release-111 (2015-07-15 10:00 UTC)
+* [CONCF-857](https://wikia-inc.atlassian.net/browse/CONCF-857) CONCF-857 Add asyncArticle to curated content routes
+* [INT-143](https://wikia-inc.atlassian.net/browse/INT-143) INT-143 Fix bug by properly using cleanTitle instead of title property
+* [HG-809](https://wikia-inc.atlassian.net/browse/HG-809) HG-809 - revison assets in all .hbs files
+* [HG-691](https://wikia-inc.atlassian.net/browse/HG-691) HG-691 setting option to get locale folders names as all lower case
+* fix-for-soc-1017: Use M.provide instead of creating new property on Mercury object
+* [SOC-1017](https://wikia-inc.atlassian.net/browse/SOC-1017) SOC-1017 Enable /both/ as default trackingMethod setting for New Auth
+
+## release-110 (2015-07-14 10:00 UTC)
+* [INT-102](https://wikia-inc.atlassian.net/browse/INT-102) Add photo to Japanese wikia
+* [INT-140](https://wikia-inc.atlassian.net/browse/INT-140) Edit/Photo icons positioning fix for section header.
+* [INT-117](https://wikia-inc.atlassian.net/browse/INT-117) Photo/Edit icons style change for Japanese Wikia
+
+## release-109 (2015-07-13 11:54 UTC)
+* [CONCF-852](https://wikia-inc.atlassian.net/browse/CONCF-852) CONCF-852 Stop failing when FeaturedContentMixin has undefined model
+
+## release-108 (2015-07-11 00:05 UTC)
+* [HG-721](https://wikia-inc.atlassian.net/browse/HG-721) HG-721 - WIP - lazy load all js
+* remove-obsolete-todo: CONCF-735 Remove obsolete TODO
+* [CONCF-735](https://wikia-inc.atlassian.net/browse/CONCF-735) CONCF-735 | Handle redirections from MercuryApi
+* [CONCF-796](https://wikia-inc.atlassian.net/browse/CONCF-796) CONCF-796 Unit tests for Curated Main Pages
+
+## release-107 (2015-07-10 10:00 UTC)
+* bucky-query-param: fixing bucky sampling query param
+* [SOC-898](https://wikia-inc.atlassian.net/browse/SOC-898) SOC-898 New Registration in wiki language
+
+## release-106 (2015-07-09 10:00 UTC)
+* [INT-104](https://wikia-inc.atlassian.net/browse/INT-104) Move section zero edit button to new block
+* [INT-130](https://wikia-inc.atlassian.net/browse/INT-130) Adding country parameter in context for performance data. 
+* [DAT-2952](https://wikia-inc.atlassian.net/browse/DAT-2952) DAT-2952 Infobox image cropping
+
+## release-105 hotfix 1 (2015-07-08 10:46 UTC)
+* [HG-753](https://wikia-inc.atlassian.net/browse/HG-753) HG-753 Add asyncArticle to the main page response
+
+## release-105 (2015-07-08 10:00 UTC)
+* [HG-753](https://wikia-inc.atlassian.net/browse/HG-753) Hg 753 - Adding ability to load first article async
+* [SOC-1005](https://wikia-inc.atlassian.net/browse/SOC-1005) SOC-1005 Trigger number keyboard when b-date input focused
+* [CONCF-655](https://wikia-inc.atlassian.net/browse/CONCF-655) CONCF-655 CuratedContent as a regular page
+* [CONCF-591](https://wikia-inc.atlassian.net/browse/CONCF-591) CONCF-591 | add load more button for categories
+
+## release-104 (2015-07-07 10:50 UTC)
+* enable-new-signup: Enabling New Registration MVP
+* [SOC-997](https://wikia-inc.atlassian.net/browse/SOC-997) SOC-997 Tracking for registration page
+* [DAT-2971](https://wikia-inc.atlassian.net/browse/DAT-2971) DAT-2971 Transitory styles for changes on app side
+* [CONCF-768](https://wikia-inc.atlassian.net/browse/CONCF-768) CONCF-768 add optimizely events for featured content
+
+## release-103 (2015-07-06 10:00 UTC)
+* [CONCF-681](https://wikia-inc.atlassian.net/browse/CONCF-681) CONCF-681 | truncating displayed labels in Curated Content and Featured Content
+
+## release-102 (2015-07-03 10:00 UTC)
+* fix-typos-on-dev-branch: Fix small typos on dev
+
+## release-101 (2015-07-02 10:34 UTC)
+* failed-unit-tests-fix: Fixing unit tests
+* [DAT-2899](https://wikia-inc.atlassian.net/browse/DAT-2899) DAT-2899 refactor icon handling in infoboxes
+* fix-infobox-gallery-button: Infoboxes- don't handle infobox collapsing when clicked on button
+* [SOC-965](https://wikia-inc.atlassian.net/browse/SOC-965) SOC-965 Rename login/signup to signin/register
+* [SOC-563](https://wikia-inc.atlassian.net/browse/SOC-563) SOC-563 Geo-targeting capabilities for New Auth
+
+## release-100 (2015-07-01 10:00 UTC)
+* [HG-782](https://wikia-inc.atlassian.net/browse/HG-782) HG-782 - added state check back
+* [HG-772](https://wikia-inc.atlassian.net/browse/HG-772) HG-772 - fixed table bug
+* [HG-777](https://wikia-inc.atlassian.net/browse/HG-777) HG-777 adding geo cookie reading
+* [SOC-698](https://wikia-inc.atlassian.net/browse/SOC-698) SOC-698 AJAX login initial commit
+* [SOC-917](https://wikia-inc.atlassian.net/browse/SOC-917) SOC-917 | Registration - login and redirect after registration
+* [SOC-913](https://wikia-inc.atlassian.net/browse/SOC-913) SOC-913 | Retire wikicitiesUserID
+
+## release-99 (2015-06-30 09:59 UTC)
+* [HG-757](https://wikia-inc.atlassian.net/browse/HG-757) HG-757 - slight optimization in replaceMediaPlaceHoldersWithMediaComponents
+* clean-ab-test: HG-727 removing ab test code
+* [CONCF-711](https://wikia-inc.atlassian.net/browse/CONCF-711) CONCF-711 | As a Mercury user I would like to have main page's URLs cut after / so I can access them easily
+* edit-api-url: Make use of buildUrl when constructing API calls
+* portable-infobox-component: Portable infobox component
+* [CONCF-676](https://wikia-inc.atlassian.net/browse/CONCF-676) CONCF-676 | As a developer I would like to have clean code related with videos in Mercury so I can add changes faster
+
+## release-98 (2015-06-29 09:59 UTC)
+* fix-helios-url-for-reg: Change helios url for reg
+* [SOC-517](https://wikia-inc.atlassian.net/browse/SOC-517) SOC-517 | Regitration error translations
+* [HG-688](https://wikia-inc.atlassian.net/browse/HG-688) HG-688 - Rendering twice when changing routes
+* [SOC-515](https://wikia-inc.atlassian.net/browse/SOC-515) Soc 515 | Registration - username and password validation errors
+
+## release-97 (2015-06-26 09:59 UTC)
+* [SOC-934](https://wikia-inc.atlassian.net/browse/SOC-934) SOC-934 Removing dice icon for Signup MVP
+* [CONCF-682](https://wikia-inc.atlassian.net/browse/CONCF-682) CONCF-682 | Internalization update
+* [CONCF-607](https://wikia-inc.atlassian.net/browse/CONCF-607) CONCF-607 Show items from all namespaces in Curated Content
+* [SOC-879](https://wikia-inc.atlassian.net/browse/SOC-879) As a PO, I want to know what errors visitors get on the registration form
+
+## release-96 (2015-06-25 09:59 UTC)
+* [SOC-895](https://wikia-inc.atlassian.net/browse/SOC-895) SOC-896 | Auth pages - Don't allow redirects to external domains
+* [SOC-755](https://wikia-inc.atlassian.net/browse/SOC-755) SOC-753 && SOC-755 Tracking for New Login flow
+* [SOC-896](https://wikia-inc.atlassian.net/browse/SOC-896) SOC-896 | Auth pages - Don't allow redirects to external domains
+
+## release-95 (2015-06-24 12:38 UTC)
+* [HG-734](https://wikia-inc.atlassian.net/browse/HG-734) HG-734 Replace assets path for ember-main.hbs only
+* fix-login-view-context: Always obtain LoginViewContext for /login
+
+## release-94 (2015-06-24 10:45 UTC)
+* [HG-734](https://wikia-inc.atlassian.net/browse/HG-734) Replace /front string in templates with cdn path
+* [HG-768](https://wikia-inc.atlassian.net/browse/HG-768) HG-768 adding query string param for bucky sampling
+* [SOC-895](https://wikia-inc.atlassian.net/browse/SOC-895) SOC-895 | Introduce AuthView + canonicalUrl for Auth pages
+* task-update: Task update
+* [SOC-798](https://wikia-inc.atlassian.net/browse/SOC-798) SOC-798 | Introduce Helios authorization
+
 ## release-93 (2015-06-23 10:59 UTC)
 * [CONCF-674](https://wikia-inc.atlassian.net/browse/CONCF-674) CONCF-674 Open videos in lightbox
 * [SOC-805](https://wikia-inc.atlassian.net/browse/SOC-805) SOC-805 changes for user login url after reverting the first merge